Player info in side bar & Enemy now move primitivly firstly

+++ src/FieldCell.cpp
@@ -9,23 +9,28 @@ FieldCell::FieldCell(unsigned ID, unsigned X, unsigned Y, bool slow, bool avaibl
     this->avaible = avaible;
 }
 
+
 unsigned FieldCell::getID() const {
     return objectID;
 }
+
 
 std::pair<int, int> FieldCell::getCoord() const {
     return {X,Y};
 }
 
+
 void FieldCell::setID(unsigned ID) {
     objectID = ID;
 }
+
 
 void FieldCell::swap(FieldCell &cell) {
     unsigned temp = cell.getID();
     cell.setID(objectID);
     this->setID(temp);
 }
+
 
 bool FieldCell::isCellSlow() const {
     return slow;
@@ -37,17 +42,19 @@ void FieldCell::setSlow(bool slowState) {
 }
 
 
-int FieldCell::getDistance(FieldCell oth) const {
+float FieldCell::getDistance(FieldCell oth) const {
     auto othCoord = oth.getCoord();
     
-    int distance = pow((std::pow((std::abs(othCoord.first - X)), 2) + std::pow((std::abs(othCoord.second - Y)), 2)), 0.5);
+    float distance = pow((std::pow((std::abs(othCoord.first - X)), 2) + std::pow((std::abs(othCoord.second - Y)), 2)), 0.5);
 
     return distance;
 }
 
+
 bool FieldCell::isCellAvaible() const {
     return avaible;
 }
+
 
 void FieldCell::setAvaible(bool avaible) {
     this->avaible = avaible;
